+++ tests/test_services.py
@@ -3,16 +3,12 @@ import json
 import pytest
 
 def test_services_influxdb(services_resource):
-    services = services_resource.get_service_info()
+    services = services_resource.get_service_list()
 
-    if 'influxdb' in services:
-        assert 'username' in services['influxdb'][0]
-        assert 'password' in services['influxdb'][0]
-        assert 'database' in services['influxdb'][0]
-        assert 'host' in services['influxdb'][0]
+    assert len(services) != 0
 
 def test_services_specific_key(services_resource):
-    services = services_resource.get_service_info(specific_key='ba96b1bebce7f02115ffe9cec442e5lb')
+    services = services_resource.get_service_info('influxdb')
 
     assert 'username' in services
     assert 'password' in services
@@ -22,5 +18,6 @@ def test_services_specific_key(services_resource):
 def test_services_key_not_exist(services_resource):
 
     with pytest.raises(Exception):
-        pytest.raises(services_resource.get_service_info(specific_key='not exist services'))
+        pytest.raises(services_resource.get_service_info('influxdb'))
 
+    assert len(services_resource.get_service_info('influxdb', service_key='123')) == 0