+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-1.2.26 (unreleased)
+1.2.26 (2018-10-26)
 -------------------
+
+- Fix bug - update services module method
+
+- Fix bug - update services module method
 
 - Add Feature - Integrate AFS hidden space